[libc++] Move _ITER_TRAITS and _ITER_CONCEPT into <__iterator/concepts.h>
